WIP on BFS for getting referenced objects
The basic code for traversing the reference chain backwards is done, but
untested. Also see all the TODOs, of which there are many, but they
don't need to be completed to determine the basic cost of traversing the
reference chain server side vs. forcing the user to traverse it in the
opposite direction client side.